requirements.txt
numpy
Pillow
moviepy==1.0.3
scipy
opencv-python-headless
